Additonal safeguard : Prefer the use of unsigned pointer sizes for reprensenting length in core functions instead of signed int in order to prevent possibly existing overflows for crafted requests. This should also fix buffer overflows in the case someone run that software on a 16 bit x86 with an already supported ᴏꜱ.

+++ src/src/base64.c
@@ -238,7 +238,7 @@ static uschar *enc64table =
   US"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
 uschar *
-b64encode(uschar *clear, int len)
+b64encode(uschar *clear, size_t len)
 {
 uschar *code = store_get(4*((len+2)/3) + 1);
 uschar *p = code;
