+++ src/src/lss.c
@@ -115,7 +115,7 @@ Returns:      a pointer to the zero-terminated base 64 string, which
 */
 
 uschar *
-lss_b64encode(uschar *clear, int len)
+lss_b64encode(uschar *clear, size_t len)
 {
 return b64encode(clear, len);
 }
